+++ 02-bpmn/src/main/scala/camundala/bpmn/BpmnProcessDsl.scala
@@ -7,23 +7,32 @@ import scala.reflect.ClassTag
 trait BpmnProcessDsl extends BpmnDsl:
 
   def processName: String
+  def processLabels: ProcessLabels = ProcessLabels.none
 
   def process[
       In <: Product: InOutEncoder: InOutDecoder: Schema,
       Out <: Product: InOutEncoder: InOutDecoder: Schema
   ](
       in: In = NoInput(),
-      out: Out = NoOutput(),
+      out: Out = NoOutput()
   ): Process[In, Out] =
-    Process(InOutDescr(processName, in, out, Some(description)))
+    Process(InOutDescr(processName, in, out, Some(description)), processLabels)
 
   private lazy val description: String =
-    s"""|$descr
+    s"""|
+        |$descr
+        |
+        |---
         |
         |- **Called Element**: `$processName` (to define in the Call Activity)
         |
+        |$processLabels.print
+        |
+        |---
+        |
+        |$companyDescr
         |""".stripMargin
-  
+
   // Use result strategy, like _singleEntry_, _collectEntries_, _singleResult_, _resultList_
   private def dmn[
       In <: Product: InOutEncoder: InOutDecoder: Schema,
@@ -38,6 +47,7 @@ trait BpmnProcessDsl extends BpmnDsl:
       InOutDescr(decisionDefinitionKey, in, out, descr.value)
     )
 
+  @deprecated("Use .. extends BpmnDecisionDsl")
   def singleEntry[
       In <: Product: InOutEncoder: InOutDecoder: Schema,
       Out <: DmnValueType: InOutEncoder: InOutDecoder: Schema: ClassTag
@@ -52,7 +62,8 @@ trait BpmnProcessDsl extends BpmnDsl:
       "A singleEntry must look like `case class SingleEntry(result: DmnValueType)`"
     ) */
     dmn(decisionDefinitionKey, in, SingleEntry(out), descr.value)
-
+    
+  @deprecated("Use .. extends BpmnDecisionDsl")
   def collectEntries[
       In <: Product: InOutEncoder: InOutDecoder: Schema,
       Out <: DmnValueType: InOutEncoder: InOutDecoder: Schema
@@ -72,6 +83,7 @@ trait BpmnProcessDsl extends BpmnDsl:
       : Conversion[Seq[Out], CollectEntries[Out]] =
     CollectEntries(_)
 
+  @deprecated("Use .. extends BpmnDecisionDsl")
   def singleResult[
       In <: Product: InOutEncoder: InOutDecoder: Schema,
       Out <: Product: InOutEncoder: InOutDecoder: Schema
@@ -90,6 +102,7 @@ trait BpmnProcessDsl extends BpmnDsl:
     ) */
     dmn(decisionDefinitionKey, in, SingleResult(out), descr.value)
 
+  @deprecated("Use .. extends BpmnDecisionDsl")
   def resultList[
       In <: Product: InOutEncoder: InOutDecoder: Schema,
       Out <: Product: InOutEncoder: InOutDecoder: Schema
@@ -108,6 +121,7 @@ trait BpmnProcessDsl extends BpmnDsl:
     )*/
     dmn(decisionDefinitionKey, in, ResultList(out), descr.value)
 
+  @deprecated("Use .. extends BpmnUserTaskDsl")
   def userTask[
       In <: Product: InOutEncoder: InOutDecoder: Schema,
       Out <: Product: InOutEncoder: InOutDecoder: Schema
@@ -121,6 +135,7 @@ trait BpmnProcessDsl extends BpmnDsl:
       InOutDescr(id, in, out, descr.value)
     )
 
+  @deprecated("Use .. extends BpmnMessageEventDsl")
   def messageEvent[
       Msg <: Product: InOutEncoder: InOutDecoder: Schema
   ](
@@ -139,7 +154,7 @@ trait BpmnProcessDsl extends BpmnDsl:
       )
     )
 
-  @deprecated("Use messageEvent.")
+  @deprecated("Use .. extends BpmnMessageEventDsl")
   def receiveMessageEvent[
       Msg <: Product: InOutEncoder: InOutDecoder: Schema
   ](
@@ -150,6 +165,7 @@ trait BpmnProcessDsl extends BpmnDsl:
   ): MessageEvent[Msg] =
     messageEvent(messageName, in, id, descr)
 
+  @deprecated("Use .. extends BpmnSignalEventDsl")
   def signalEvent[
       Msg <: Product: InOutEncoder: InOutDecoder: Schema
   ](
@@ -168,7 +184,7 @@ trait BpmnProcessDsl extends BpmnDsl:
       )
     )
 
-  @deprecated("Use signalEvent.")
+  @deprecated("Use .. extends BpmnSignalEventDsl")
   def receiveSignalEvent[
       Msg <: Product: InOutEncoder: InOutDecoder: Schema
   ](
@@ -179,6 +195,7 @@ trait BpmnProcessDsl extends BpmnDsl:
   ): SignalEvent[Msg] =
     signalEvent(messageName, in, id, descr)
 
+  @deprecated("Use .. extends BpmnTimerDsl")
   def timerEvent(
       title: String,
       descr: Optable[String] = None