+++ 04-helper/src/main/scala/camundala/helper/DevHelper.scala
@@ -18,11 +18,17 @@ object DevHelper:
     SetupGenerator()(using config.setupConfig).generate
 
   // project (./projects/mycompany-myproject/helper.sc)
-  def update(config: HelperConfig, subProjects: Seq[String]): Unit =
-    given SetupConfig = config.setupConfig
-      .copy(subProjects = subProjects)
-    println(s"Update Project: ${config.setupConfig.projectName}")
-    println(s" - SubProjects: $subProjects")
+  def update(using config: SetupConfig): Unit =
+    println(s"Update Project: ${config.projectName}")
+    println(s" - with Subprojects: ${config.subProjects}")
     SetupGenerator().generate
+    
+  def createProcess(processName: String)(using config: SetupConfig): Unit =
+    println(s"Create Process: $processName in ${config.projectName}")
+    SetupGenerator().createProcess(processName)
+  
+  def createCustomWorker(processName: String, workerName: String)(using config: SetupConfig): Unit =
+    println(s"Create Custom Worker: $workerName in ${config.projectName} / process: $processName")
+    SetupGenerator().createCustomWorker(processName, workerName)
 
 end DevHelper