+++ 04-helper/src/main/scala/camundala/helper/setup/ScriptCreator.scala
@@ -1,16 +1,12 @@
 package camundala.helper.setup
 
-import camundala.helper.util.{CompanyVersionHelper, ReposConfig}
+import camundala.helper.util.CompanyVersionHelper
 
 case class ScriptCreator()(using config: SetupConfig):
 
   lazy val companyCreate =
     s"""$helperHeader
        |
-       |
-       |/**
-       | * Usage see `camundala.helper.DevHelper.createUpdateCompany`
-       | */
        |@main(doc =
        |  \"\"\"> Creates the directories and generic files for the company BPMN Projects
        |   \"\"\")
@@ -23,9 +19,6 @@ case class ScriptCreator()(using config: SetupConfig):
   lazy val projectCreate: String =
     s"""$helperHeader
        |
-       |/**
-       | * Usage see `$companyName.camundala.helper.SetupHelper#createCompany`
-       | */
        |@main(doc =
        |  \"\"\"> Creates the directories and generic files for the company BPMN Projects
        |   \"\"\")
@@ -42,9 +35,34 @@ case class ScriptCreator()(using config: SetupConfig):
     val projectName = config.projectName
     s"""$helperHeader
        |
-       |/**
-       | * Usage see `valiant.camundala.helper.PublishHelper`
-       | */
+       |lazy val projectName: String = "$projectName"
+       |lazy val subProjects = Seq(
+       |  ${config.subProjects.map(sp => s"\"$sp\"").mkString(", ")}
+       |)
+       |lazy val config: HelperConfig = ProjectDevHelper.config(projectName, subProjects)
+       |given setup.SetupConfig =  config.setupConfig
+       |
+       |@main(doc =
+       |  \"\"\"> Updates your Project with latest versions and also updates generic files, that starts with '$doNotAdjust'.
+       |      - set in `helper.sc` the version you want: $helperImport
+       |   \"\"\")
+       |def update(): Unit = DevHelper.update
+       |
+       |@main(doc = "> Creates everything for a Process (bpmn, simulation, worker)")
+       |def process(
+       |             @arg(doc = "The name of the process (lowerCase, e.g. mySuperProcess).")
+       |             processName: String,
+       |             ): Unit =
+       |  DevHelper.createProcess(processName)
+       |
+       |@main(doc = "> Creates everything for a CustomWorker (bpmn, simulation, worker)")
+       |def customWorker(
+       |             @arg(doc = "The name of the process.")
+       |             processName: String,
+       |             @arg(doc = "The name of the Worker (without 'Worker').")
+       |             workerName: String,
+       |             ): Unit =
+       |  DevHelper.createCustomWorker(processName, workerName)
        |
        |@main(doc = "> Creates a new Release for the client and publishes to bpf-generic-release")
        |def publish(
@@ -53,9 +71,6 @@ case class ScriptCreator()(using config: SetupConfig):
        |  PublishHelper().publish(version)
        |}
        |
-       |/**
-       | * Usage see `valiant.camundala.helper.DeployHelper`
-       | */
        |@main(doc = "> Publishes Local (sbt publishLocal) and runs the deploy_manifest of Postman")
        |def deploy(
        |            @arg(doc = "Optional Gatling Test that is run after deployment.")
@@ -63,47 +78,23 @@ case class ScriptCreator()(using config: SetupConfig):
        |  DeployHelper().deploy(integrationTest)
        |}
        |
-       |/**
-       | * Usage see `valiant.camundala.helper.UpdateHelper`
-       | */
-       |@main(doc =
-       |  \"\"\"> Updates your Project with latest versions and also updates generic files, that starts with '$doNotAdjust'.
-       |      - set in `helper.sc` the version you want: $helperImport
-       |   \"\"\")
-       |def update(
-       |    @arg(doc = "The project name - should be generated automatically after creation.")
-       |    projectName: String = "$projectName",
-       |): Unit =
-       |  val config = ProjectDevHelper.config(projectName)
-       |    val subProjects = Seq(
-       |      ${config.subProjects.map(sp => s"\"$sp\"").mkString(", ") }
-       |    )
-       |  DevHelper.update(config, subProjects)
        |
-       |/**
-       | * Usage see `valiant.camundala.helper.DockerHelper`
-       | */
        |@main(doc = "> Starts your development Docker - this assumes the Docker Files in `../../docker`.")
        |def dockerUp(): Unit = {
        |  DockerHelper().dockerUp()
        |}
        |
-       |/**
-       | * Usage see `valiant.camundala.helper.DockerHelper`
-       | */
        |@main(doc = "> Stops your development Docker - this assumes the Docker Files in `../../docker`.")
        |def dockerStop(): Unit = {
        |  DockerHelper().dockerStop()
        |}
        |
-       |/**
-       | * Usage see `valiant.camundala.helper.DockerHelper`
-       | */
        |@main(doc = "> Stops and removes your development Docker - this assumes the Docker Files in `../../docker`.")
        |def dockerDown(): Unit = {
        |  DockerHelper().dockerDown()
        |}
        |""".stripMargin
+  end projectHelper
 
   private lazy val companyName = config.companyName
   private lazy val reposConfig = config.reposConfig
@@ -122,10 +113,11 @@ case class ScriptCreator()(using config: SetupConfig):
        |
        |interp.repositories() ++= Seq(
        |  ${
-      reposConfig.ammoniteRepos
-        .map:
-          _.ammoniteRepo
-        .mkString(",\n  ")}
+        reposConfig.ammoniteRepos
+          .map:
+            _.ammoniteRepo
+          .mkString(",\n  ")
+      }
        |)
        |@
        |