+++ 04-helper/src/main/scala/camundala/helper/setup/SetupGenerator.scala
@@ -10,4 +10,14 @@ case class SetupGenerator()(using config: SetupConfig):
     WorkerGenerator().generate
     HelperGenerator().generate
   end generate
+  
+  def createProcess(processName: String): Unit =
+    BpmnGenerator().createProcess(processName)
+    SimulationGenerator().createProcess(processName)
+    WorkerGenerator().createProcess(processName)
+    
+  def createCustomWorker(processName: String, workerName: String): Unit =
+    BpmnGenerator().createCustomWorker(processName, workerName)
+    WorkerGenerator().createCustomWorker(processName, workerName)
+
 end SetupGenerator