+++ 04-worker-c7spring/src/main/scala/camundala/worker/springDeps.scala
@@ -1,0 +1,9 @@
+package camundala.worker
+
+// Worker
+export org.springframework.context.annotation.Configuration
+// WorkerApp
+export org.springframework.boot.SpringApplication
+export org.springframework.boot.autoconfigure.SpringBootApplication
+export org.springframework.boot.context.properties.ConfigurationPropertiesScan
+export org.springframework.context.annotation.ComponentScan