+++ 05-examples/demos/02-bpmn/src/main/scala/camundala/examples/demos/bpmn/SignalMessageExample.scala
@@ -6,6 +6,7 @@ import camundala.domain.*
 object MessageForExample extends BpmnProcessDsl:
   val processName = "message-for-example"
   val descr = ""
+  val companyDescr = ""
   lazy val messageExample = process(
     in = SignalMessageExampleIn(),
     out = SignalMessageExampleOut()
@@ -19,6 +20,7 @@ end MessageForExample
 object SignalExample extends BpmnProcessDsl:
   val processName = "signal-example"
   val descr = ""
+  val companyDescr = ""
 
   lazy val signalExample = process(
     in = SignalMessageExampleIn(),