+++ 05-examples/demos/02-bpmn/src/main/scala/camundala/examples/demos/bpmn/TimerExample.scala
@@ -6,6 +6,7 @@ object TimerExample extends BpmnProcessDsl:
 
   lazy val processName: String = "timer-example"
   lazy val descr: String = ""
+  val companyDescr = ""
 
   lazy val timerProcess = process()